Move assertion checks into conftest
